Helpers: keep the css input types where both wrappers can reach them
